implementing wasm code converter
