+++ cmdexec/js.wasm.go
@@ -6,15 +6,26 @@ import (
 	"syscall/js"
 
 	"github.com/dop251/goja"
+	"github.com/dop251/goja_nodejs/console"
+	"github.com/dop251/goja_nodejs/require"
 )
 
 func RunVMCode(jsCode string, vm *goja.Runtime) {
 	js.Global().Call("eval", jsCode)
 }
 
-//  The GenerateChainCode function creates a JavaScript code snippet
-//  that attempts to parse a given HTTP response body as JSON, logging
-//  success or failure, and stores the result accordingly.
+// GetJSVm creates a new goja runtime instance
+// with console.log enabled
+func GetJSVm() *goja.Runtime {
+	vm := goja.New()
+	new(require.Registry).Enable(vm)
+	console.Enable(vm)
+	return vm
+}
+
+// The GenerateChainCode function creates a JavaScript code snippet
+// that attempts to parse a given HTTP response body as JSON, logging
+// success or failure, and stores the result accordingly.
 func GenerateChainCode(httpRespBody string) string {
 	code := `try {
 		result = JSON.parse(String.raw` + "`" + httpRespBody + "`" + `)
